@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-ruby '2.7.8'
+ruby '3.2.4'
 gem 'rails', '6.1.7'
 
 # AUTH
